+++ rtc_base/rate_tracker.h
@@ -21,7 +21,7 @@ namespace rtc {
 // that over each bucket the rate was constant.
 class RateTracker {
  public:
-  RateTracker(int64_t bucket_milliseconds, size_t bucket_count);
+  RateTracker(int64_t bucket_milliseconds, uint64_t bucket_count);
   virtual ~RateTracker();
 
   // Computes the average rate over the most recent interval_milliseconds,
@@ -41,11 +41,11 @@ class RateTracker {
   double ComputeTotalRate() const;
 
   // The total number of samples added.
-  size_t TotalSampleCount() const;
+  uint64_t TotalSampleCount() const;
 
   // Reads the current time in order to determine the appropriate bucket for
   // these samples, and increments the count for that bucket by sample_count.
-  void AddSamples(size_t sample_count);
+  void AddSamples(uint64_t sample_count);
 
  protected:
   // overrideable for tests
@@ -56,8 +56,8 @@ class RateTracker {
   size_t NextBucketIndex(size_t bucket_index) const;
 
   const int64_t bucket_milliseconds_;
-  const size_t bucket_count_;
-  size_t* sample_buckets_;
+  const uint64_t bucket_count_;
+  uint64_t* sample_buckets_;
   size_t total_sample_count_;
   size_t current_bucket_;
   int64_t bucket_start_time_milliseconds_;
